Fixes Richards benchmark and forgot to include file in previous CL for dynamic.
BUG=
R=andrewkrieger@google.com

Review URL: https://chromereviews.googleplex.com//495677014 .

+++ test/unit/dynamic/scenario.dart
@@ -13,8 +13,8 @@ class Class2 {
 int testDynamicDispatch() {
   int result = 0;
   dynamic d = new Class1();
-  result = result + d.foo(10, 20);
+  result = result + (d.foo(10, 20) as int);
   d = new Class2();
-  result = result + d.foo(100, 200);
+  result = result + (d.foo(100, 200) as int);
   return result;
 }
